pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aurora-bmi"
version = "1.0.0"
description = "AURORA BMI - Baseline-normalized Market Breadth Index"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AURORA Team"},
]
keywords = ["market-breadth", "quantitative-finance", "market-diagnostics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # HTTP & async
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",

    # Configuration
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",

    # Dashboard
    "streamlit>=1.31.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "aioresponses>=0.7.6",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pandas-stubs>=2.0.0",
]

[project.scripts]
aurora-daily = "scripts.run_daily:main"
aurora-diagnose = "scripts.diagnose_api:main"

[tool.hatch.build.targets.wheel]
packages = ["aurora"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may use I/O)",
    "slow: Tests that take >1 second",
    "api: Tests that require API keys",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
